Move everything around
The new middleware isn't auth0 specific, so this moves everything around to
prepare for that.
