+++ mozilla_django_oidc/middleware.py
@@ -1,9 +1,14 @@
+try:
+    from urllib import urlencode
+except ImportError:
+    from urllib.parse import urlencode
+
+import requests
+from six import string_types
+
 import django
 from django.core.cache import cache
 
-from six import string_types
-
-from mozilla_django_oidc.contrib.auth0.utils import refresh_id_token
 from mozilla_django_oidc.utils import import_from_settings, is_authenticated
 from mozilla_django_oidc.views import OIDCLogoutView
 
@@ -15,6 +20,42 @@ if tuple(django.VERSION[0:2]) >= (1, 10):
 else:
     class MiddlewareMixin(object):
         pass
+
+
+def refresh_id_token(id_token):
+    """Renews the id_token from the delegation endpoint in Auth0.
+
+    :arg str id_token: the id token to renew
+
+    :returns: ``id_token`` if renewed successfully or ``None``
+
+    """
+    # FIXME(willkg): Rewrite this to use authorization endpoint
+    delegation_url = 'https://{0}/delegation'.format(import_from_settings('OIDC_OP_DOMAIN'))
+    data = {
+        'client_id': import_from_settings('OIDC_RP_CLIENT_ID'),
+        'grant_type': 'urn:ietf:params:oauth:grant-type:jwt-bearer',
+        'id_token': id_token,
+        'api_type': 'app'
+    }
+
+    response = requests.post(delegation_url, data=data)
+
+    if response.status_code == requests.codes.ok:
+        return response.json().get('id_token')
+
+    return
+
+
+def logout_url():
+    """Log out the user from Auth0."""
+    # FIXME(willkg): This doesn't appear to be used anywhere.
+    url = 'https://' + import_from_settings('OIDC_OP_DOMAIN') + '/v2/logout'
+    url += '?' + urlencode({
+        'returnTo': import_from_settings('LOGOUT_REDIRECT_URL', '/'),
+        'client_id': import_from_settings('OIDC_RP_CLIENT_ID')
+    })
+    return url
 
 
 class RefreshIDToken(MiddlewareMixin):
